Blender Addon: Speed up loading FA dumps when multiple files of the same draw call are selected
When multiple files of the same draw call are selected during import
they will all trigger the code to look for the other matching files from
their draw call, which of course means the results of each of these will
be identical and the duplicates are eliminated when they get added to
the ret set.

This adds the related files found during the glob to a second set that
is used to skip searching for their related files before the glob, which
should help save some time when importing large frame analysis folders.
In the Dreamfall dump mentioned in the previous commit it doesn't save
much, only about 200-300ms, but that was a relatively small dump
captured with very little on screen.