Blender Addon: Improve handling of importing unsupported files
If the user attempts to import all files in a frame analysis directory
with 'load related' disabled they will get a fatal error "Unable to find
corresponding buffers from filename..." if any of the selected files are
not supported for import, such as the frame analysis log.txt file,
ShaderUsage.txt, any constant buffers dumped with dump_cb, or any other
buffer dumped that is not a vertex or index buffer.

This commit addresses these issues in a few ways:
- Include the filename in the error message so the user can work out why
  they are seeing the error.
- Don't treat it as fatal, continue with other files in the import
- Skip log.txt and ShaderUsage.txt without triggering the error
- Change note on error message to be more specific about what the
  problem is rather than theorising on a possible (and incorrect) cause.
  I don't think anyone has tried loading from a deduped folder anyway.

Note that doing this procedure with "Load related" enabled will not
trigger this error, as in that case the file list is replaced by a list
of files that match the vertex or index buffer hash, which will not be
present in the problematic files.

Possibly an alternate explanation for the confusion here:
https://github.com/bo3b/3Dmigoto/issues/151