Blender Addon: Add support for Blender 4.0
Blender 4.0 dropped vertex layers in favour of a more flexible custom
attributes API. This adds compatibility accessors for these layers /
attributes to work with both Blender 3.x and 4.x.

Currently the attributes are handled in exactly the same way as the
vertex layers - there is some potential to do a better job on 4.x, for
example 2 or 3 component float semantics could be stored as the FLOAT2 /
FLOAT_VECTOR types instead of 2 or 3 individual attributes. However,
despite the increased flexibility, Blender is still lacking a 4
component float type, which is arguably the most common one we would
want (as 4 is the number of components in a typical GPU register), so we
would still end up needing to break up those semantics anyway... and at
that point it's probably not worth the added complexity just to achieve
a half-arsed result.

Currently migrating between 3.x and 4.x doesn't really work - best to
just re-import the frame analysis or raw buffers on 4.x rather than
trying to open a .blend file saved in 3.x. It looks like custom
attributes were added early in 3.x, so we might later be able to add a
migration operator that could be run on 3.x to convert.

Closes https://github.com/DarkStarSword/3d-fixes/issues/24