Blender Addon: Add Frame Analysis log.txt parser
This adds a class that can parse frame analysis log files, keeping track
of bound resource states (including subsequent draw calls that don't
change the bindings) and allowing queries for where resources have been
used.

This is not currently hooked up to anything, but it is intended to be
used to find vertex buffers that have been written to by Stream Output
geometry shaders in skinning passes in e.g. Unity games, so that we can
locate the original unposed vertex buffers when importing to Blender.

I considered a simpler implementation that would solve only the above
problem more concisely, but ended up deciding on this implementation
that can easily be extended to track more state from the log file if we
need it in the future - there are multiple commented out (to minimise
unnecessary overhead) classes that can by used to easily add tracking
for additional resource slots, and it should be similarly straight
forward to add tracking for other state (such as bound shaders).

This should be one step towards solving:
https://github.com/DarkStarSword/3d-fixes/issues/19