Blender Addon: WIP loading vertex buffers from Stream Output passes
Adds a hidden (because not yet working) option to use the frame analysis
log file parser to locate any stream output passes associated with the
vertex buffers being loaded and adds the vertex buffers from those SO
passes as extra buffers to be loaded.

Still needs work to properly handle the extra buffers, but this at least
allows it to find which buffers will need to be loaded.

Working-Towards: https://github.com/DarkStarSword/3d-fixes/issues/19