Blender Addon: Fix use of wrong "operator" when prefilling semantics
This was picking up "operator" from the module wide globals rather than
a local passed to the function, and just didn't matter if nothing
attempted to call operator.report().

Fix it to use the correct "operator" and move the "import operator" into
the function that needs it so it doesn't inadvertently get picked up by
anything else.