Blender Addon: Handle UNORM normals, and add flip normal option
Unreal stores normals + tangents in UNORM data types, meaning these are
unit vectors with range -1:+1 being stored in a data type with range
0:+1. This adds a conversion on import and export to rescale the range.

This also adds an option to flip the normals on import/export, which
also seemed necessary to get Unreal normals to appear correct in
Blender. However, face orientation also shows the faces as red, so I'm
now more convinced that this was really due to either the winding order
being reversed, or possibly a flipped axis, but since this support is
working and may be useful in some situations, commit it anyway.

Related-to: https://github.com/DarkStarSword/3d-fixes/issues/25