Blender Addon: Add option to reverse winding order on import/export
Unreal Engine seems to use the opposite triangle winding order to other
games, which results in the mesh displaying red if the face orientation
overlay is enabled, or the wrong faces visible if backface culling is
enabled.

Normally this sort of issue would be resolved by flipping the normals in
Blender, but that process actually does both reversing the winding order
and flipping the normals - since in this case the normals are already
correct, we just want the winding order reversed.

This adds an option to reverse the winding order during import/export to
handle these cases. The option is saved in the object custom data so the
export option will automatically be set to match what was used during
import, but a modder can toggle it separately for export in case they
have manually flipped the normals while working on the mesh and need to
correct for that.

Related-to: https://github.com/DarkStarSword/3d-fixes/issues/25