Blender Addon: Tag object with all optional hashes from the filename regex
- If we are storing the vertex + pixel shader hash, we might as well
  store the hull, domain and geometry shader hashes if present.
- Don't store a blank hash if not present
- Long line is split up because horizontal scrolling is bad

This will also store hash contamination warnings if present since I
added a group name for them in the regex. Not sure what we want to do
with these if they are present (warn on import? warn on export? write
warning into ini file? All of the above? Just ignore them because they
are fairly rare on vertex buffers?)